refactor(commons): lift the language-neutral bounds, paging and key helpers out of the Python backend
A pure move. The validation, keyset paging, edge ordering, bounded-subgraph,
path-assembly and module-key helpers leg 1.5/1.6 built contained nothing
Python-specific except three things, now parameters: the PY_ relationship
prefix, the .py extension, and the VIA map. Python re-imports every name, so
its suite is the proof that nothing changed. TypeScript (leg 2.5) and Java
(leg 2) inherit the same rulings instead of forking them.

The one test that read DEFAULT_DEPTH's rule out of the Python backend's source
text now reads it from commons.bounds, where the constant and its rule live.
